+++ lockfile/src/parsers/mod.rs
@@ -1,9 +1,9 @@
 use nom::branch::alt;
 use nom::bytes::complete::{tag, take_until};
-use nom::character::complete::{line_ending, multispace0, not_line_ending, space0};
+use nom::character::complete::{line_ending, not_line_ending, space0};
 use nom::combinator::{eof, opt, recognize, rest};
 use nom::error::{context, VerboseError};
-use nom::multi::{count, many1, many_till};
+use nom::multi::many_till;
 use nom::sequence::{delimited, terminated, tuple};
 use nom::AsChar;